remove get buffer
